Update main.c
Made small changes for my function. Its an attempt to take the user input from resource types (if there are 4 resources then the size should be 4) and allocate a vector that will only hold the value for the quantity of a certain resource type for  a process needs to finish (i.e. Process 1 needs 5A, 2B so need [0] = 5 and need [1] = 2).